feat: first chatGPT input
